run-atm changes plus added vmd template that was missing

+++ examples/ABFE/fkbp/scripts/run_template.sh
@@ -1,14 +1,22 @@
 #!/bin/bash
 #
 #SBATCH -J <JOBNAME>
-#SBATCH --partition=<PARTITION>
+#SBATCH --partition=any
 #SBATCH --nodes=1
 #SBATCH --ntasks-per-node=1
-#SBATCH --gpus=1
-#SBATCH --account=<ACCOUNTNO>
+#SBATCH --gres=gpu:1
+#SBATCH --cpus-per-task=3
 #SBATCH --no-requeue
-#SBATCH -t 02:15:00
+#SBATCH -t 26:00:00
 
-. $HOME/miniforge3/bin/activate
+jobname=<JOBNAME>
+
+. ${HOME}/miniconda3/bin/activate atm8.4.0
+echo "Running on $(hostname)"
+
+if [ ! -f ${jobname}_0.xml ]; then
+    abfe_structprep ${jobname}_asyncre.cntl || exit 1
+fi
+
 echo "localhost,0:0,1,CUDA,,/tmp" > nodefile
-abfe_production <JOBNAME>_asyncre.cntl
+abfe_production ${jobname}_asyncre.cntl
